+++ contracts/Account.sol
@@ -8,6 +8,11 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "./IAccount.sol";
 
 contract Account is IAccount, ERC721Holder {
+    bytes32 private _slug;
+    constructor(bytes32 slug_) {
+        _slug = slug_;
+    }
+
     function approveFungibleToken(address token_, address spender_, uint256 amount_) public override {
         IERC20(token_).approve(spender_, amount_);
     }
@@ -18,5 +23,9 @@ contract Account is IAccount, ERC721Holder {
 
     function approveAllNonFungibleToken(address token_, address operator_, bool approved_) public override {
         IERC721(token_).setApprovalForAll(operator_, approved_);
+    }
+
+    function slug() external view override returns (bytes32) {
+        return _slug;
     }
 }