+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add `_createAccount`, `_deleteAccount` and `isAccount` method for AccountProvider.
+- `Account` now require slug\_ at construction and support `.slug()` method.
+
 ## 0.1.1-0 (2021-10-05)
 
 - Setup repo